fixed kill 8080

+++ codedeploy/applicationStart.sh
@@ -1,7 +1,13 @@
 #!/bin/bash -v
 
-sudo systemctl stop tomcat8
 # shellcheck disable=SC2046
-sudo kill -9 $(lsof -t -i:8080)
+if [ -z $(lsof -t -i:8080) ]
+then
+      echo "nothing runs on 8080"
+else
+      sudo kill -9 $(lsof -t -i:8080)
+fi
+
+
 
 sudo java -jar /home/webapp.jar
